+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # [Versions](https://mui.com/versions/)
 
+## 0.0.22
+
+<!-- generated comparing v0.0.21..master -->
+
+_Sep 3, 2024_
+
+A big thanks to the 3 contributors who made this release possible.
+
+### `@pigment-css/vite-plugin@0.0.22`
+
+- &#8203;<!-- 1 -->Fix bug related to path separator on Windows (#218) @brijeshb42
+- &#8203;<!-- 4 -->Fix styled-engine new internals (#220) @romgrk
+
+### `@pigment-css/unplugin@0.0.22`
+
+- &#8203;<!-- 1 -->Fix bug related to path separator on Windows (#218) @brijeshb42
+- &#8203;<!-- 4 -->Fix styled-engine new internals (#220) @romgrk
+
+### Core
+
+- &#8203;<!-- 3 -->[core] Pin GitHub action dependency (#221) @oliviertassinari
+- &#8203;<!-- 2 -->[core] Fix CI with fake docs (#222) @oliviertassinari
+
+All contributors of this release in alphabetical order: @brijeshb42, @oliviertassinari, @romgrk
+
 ## 0.0.21
 
 <!-- generated comparing v0.0.20..master -->